move old code to legacy dir
